put common helper methods in setup file
