+++ tests/setup.py
@@ -1,5 +1,7 @@
 import subprocess
 import beebop.schemas
+from beebop import assignClusters
+from beebop import visualise
 from tests import hdf5_to_json
 import json
 
@@ -19,3 +21,36 @@ def generate_json():
     sketches_json = json.loads(hdf5_to_json.h5_to_json(filepath))
 
     return json.dumps(sketches_json)
+
+expected_assign_result = {
+     0: {'cluster': 'GPSC16', 'hash': '02ff334f17f17d775b9ecd69046ed296'},
+     1: {'cluster': 'GPSC29', 'hash': '9c00583e2f24fed5e3c6baa87a4bfa4c'},
+     2: {'cluster': 'GPSC8', 'hash': '99965c83b1839b25c3c27bd2910da00a'}
+}
+
+name_mapping = {
+    "02ff334f17f17d775b9ecd69046ed296": "name1.fa",
+    "9c00583e2f24fed5e3c6baa87a4bfa4c": "name2.fa"
+}
+
+def do_assign_clusters(p_hash: str):
+    hashes_list = [
+            '02ff334f17f17d775b9ecd69046ed296',
+            '9c00583e2f24fed5e3c6baa87a4bfa4c',
+            '99965c83b1839b25c3c27bd2910da00a']
+
+    return assignClusters.get_clusters(
+        hashes_list,
+        p_hash,
+        fs,
+        db_paths,
+        args)
+
+def do_network_internal(p_hash: str):
+    do_assign_clusters(p_hash)
+    visualise.network_internal(expected_assign_result,
+                               p_hash,
+                               fs,
+                               db_paths,
+                               args,
+                               name_mapping)
