+++ tests/test_integration.py
@@ -3,7 +3,6 @@ import jsonschema
 import os
 import beebop.schemas
 from tests import setup
-from tests.test_unit import do_network_internal
 
 
 schemas = beebop.schemas.Schema()
